feat: Added additional endpoints for users

+++ database/repositories/userRepository.py
@@ -1,13 +1,8 @@
 import bcrypt
 from database.base import db
-from database.models.user import User, VALID_ROLES
+from database.models.user import User
 
 def createUser(name, email, password, role):
-    # Validates the input
-    if role not in VALID_ROLES:
-        print(f'ERROR: Role {role} is not valid')
-        return
-
     # --- Encrypt password ---
     # Adding the salt to password
     salt = bcrypt.gensalt()
@@ -25,10 +20,12 @@ def createUser(name, email, password, role):
         db.session.commit()
         print('The user was successfully created!')
     except Exception as error:
-        print(str(error.orig) + " for parameters" + str(error.params))
+        raise Exception(str(error.orig) + " for parameters" + str(error.params))
 
     # Close db.session
     db.session.close()
+
+    return
 
 def getAllUsers():
     # Get data from DB
@@ -36,7 +33,7 @@ def getAllUsers():
     try:
         users = db.session.query(User).all()
     except Exception as error:
-        print(str(error.orig) + " for parameters" + str(error.params))
+        raise Exception(str(error.orig) + " for parameters" + str(error.params))
 
     # Close db.session
     db.session.close()
@@ -44,17 +41,14 @@ def getAllUsers():
     return users
 
 def updateUser(id, name, email, password, role):
-    # Validates the input
-    if role not in VALID_ROLES:
-        print(f'ERROR: Role {role} is not valid')
-        return
-
     # Get user from DB
     try:
         user = db.session.query(User).get(id)
     except Exception as error:
-        print(str(error.orig) + " for parameters" + str(error.params))
-        return
+        raise Exception(str(error.orig) + " for parameters" + str(error.params))
+
+    if not user:
+        raise Exception(f'User with ID {id} was not found')
 
     # Update the user's info
     user.name = name
@@ -66,7 +60,7 @@ def updateUser(id, name, email, password, role):
         db.session.commit()
         print('The user was successfully updated!')
     except Exception as error:
-        print(str(error.orig) + " for parameters" + str(error.params))
+        raise Exception(str(error.orig) + " for parameters" + str(error.params))
 
     # Close db.session
     db.session.close()
@@ -76,8 +70,10 @@ def removeUser(id):
     try:
         user = db.session.query(User).get(id)
     except Exception as error:
-        print(str(error.orig) + " for parameters" + str(error.params))
-        return
+        raise Exception(str(error.orig) + " for parameters" + str(error.params))
+
+    if not user:
+        raise Exception(f'User with ID {id} was not found')
 
     # Remove the user
     db.session.delete(user)
@@ -87,7 +83,7 @@ def removeUser(id):
         db.session.commit()
         print('The user was successfully removed!')
     except Exception as error:
-        print(str(error.orig) + " for parameters" + str(error.params))
+        raise Exception(str(error.orig) + " for parameters" + str(error.params))
 
     # Close db.session
     db.session.close()
